Add async operation handler
