+++ src/ansys/hps/data_transfer/client/api/__init__.py
@@ -24,3 +24,4 @@
 
 from .api import DataTransferApi
 from .async_api import AsyncDataTransferApi
+from .handler import DefaultOperationHandler, AsyncOperationHandler
