+++ src/ansys/hps/data_transfer/client/api/async_api.py
@@ -54,6 +54,7 @@ from ..models.ops import Operation, OperationState
 from ..models.permissions import RoleAssignment, RoleQuery
 from ..utils.jitter import get_expo_backoff
 from .retry import retry
+from .handler import AsyncOperationHandler
 
 log = logging.getLogger(__name__)
 
@@ -205,7 +206,8 @@ class AsyncDataTransferApi:
         interval: float = 0.1,
         cap: float = 2.0,
         raise_on_error: bool = False,
-        progress_handler: Callable[[str, float], Awaitable[None]] = None,
+        # progress_handler: Callable[[str, float], Awaitable[None]] = None,
+        operation_handler: Callable[[builtins.list[Operation]], Awaitable[None]] = None,
     ):
         """Provides an async interface to wait for a list of operations to complete.
 
@@ -221,10 +223,12 @@ class AsyncDataTransferApi:
             The maximum backoff value used to calculate the next wait time. Default is 2.0.
         raise_on_error: bool
             Raise an exception if an error occurs. Default is False.
-        progress_handler: Callable[[str, float], None]
-            A async function to handle progress updates. Default is None.
-
+        operation_handler: Callable[[builtins.list[Operation]], None]
+            A callable that will be called with the list of operations when they are fetched.
         """
+        if operation_handler is None:
+            operation_handler = AsyncOperationHandler()
+
         if not isinstance(operation_ids, list):
             operation_ids = [operation_ids]
         operation_ids = [op.id if isinstance(op, Operation | OpIdResponse) else op for op in operation_ids]
@@ -236,22 +240,19 @@ class AsyncDataTransferApi:
             attempt += 1
             try:
                 ops = await self._operations(operation_ids)
-                so_far = hf.format_timespan(time.time() - start)
-                log.debug(f"Waiting for {len(operation_ids)} operations to complete, {so_far} so far")
-                if self.client.binary_config.debug:
+
+                meta = getattr(operation_handler, "Meta", dict())
+                expand_groups = getattr(meta, "expand_groups", False)
+                if expand_groups:
+                    expanded = []
                     for op in ops:
-                        fields = [
-                            f"id={op.id}",
-                            f"state={op.state}",
-                            f"start={op.started_at}",
-                            f"succeeded_on={op.succeeded_on}",
-                        ]
-                        if op.progress > 0:
-                            fields.append(f"progress={op.progress:.3f}")
-                        log.debug(f"- Operation '{op.description}' {' '.join(fields)}")
-                if progress_handler is not None:
-                    for op in ops:
-                        await progress_handler(op.id, op.progress)
+                        if not op.children:
+                            continue
+                        expanded.extend(await self._operations(op.children))
+                        expanded.append(op)
+                    if operation_handler is not None:
+                        await operation_handler(expanded)
+
                 if all(op.state in [OperationState.Succeeded, OperationState.Failed] for op in ops):
                     break
             except Exception as e:
